dental-clinic-backend/requirements.txt
Django==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
djoser==2.2.0
django-model-utils==4.3.1
Pillow==10.1.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.9